Fix newlines and AOSP formatting and import order
Apparently I forgot to do this with the initial commit. All files now
have the correct newlines, have been run through the AOSP code formatter
(making several things a uglier in the process) and have the correct
import order. This is very painful commit patch-wise, but it needs to be
done.